+++ packages/api/src/configuration/exceptions/exception.abstract.ts
