+++ packages/api/src/configuration/index.ts
@@ -1,5 +1,3 @@
 export const DATA_BASE_CONFIGURATION = {
   mongoConnectionString: "mongodb://localhost:27017/nest",
 };
-
-//  mongoConnectionString: process.env.CLEAN_NEST_MONGO_CONNECTION_STRING as string,
