--- packages/api/src/controllers/app.controller.ts
@@ -1,9 +1,0 @@
-import { Controller, Get } from "@nestjs/common";
-
-@Controller()
-export class AppController {
-  @Get()
-  getHello(): string {
-    return "SPAship is active";
-  }
-}
