--- packages/api/src/controllers/index.ts
@@ -1,5 +1,0 @@
-export * from "./app.controller";
-export * from "../server/application/application.controller";
-export * from "../server/deployment-record/deployment-record.controller";
-export * from "../server/api-key/apikey.controller";
-export * from "../server/deployment-conncetion/deployment-connection.controller";
