--- packages/api/src/core/abstracts/deployment-services.service.ts
@@ -1,5 +1,0 @@
-import { Application } from "../entities/application.entity";
-
-export abstract class IDeploymentServices {
-  abstract deployApplication(application: Application): Promise<Application>;
-}
