--- packages/api/src/core/abstracts/index.ts
@@ -1,3 +1,0 @@
-export * from "./data-services.abstract";
-export * from "./generic-repository.abstract";
-export * from "./deployment-services.service";
