--- packages/api/src/core/index.ts
@@ -1,2 +1,0 @@
-// export * from "./abstracts";
-// export * from "./entities";
