+++ packages/api/src/repository/data-services.module.ts
@@ -1,5 +1,5 @@
 import { Module } from "@nestjs/common";
-import { MongoDataServicesModule } from "../../repository/mongo/mongo-data-services.module";
+import { MongoDataServicesModule } from "./mongo/mongo-data-services.module";
 
 @Module({
   imports: [MongoDataServicesModule],
