+++ packages/api/src/server/application/application-response.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from "@nestjs/swagger";
-import { Application } from "src/core/entities/application.entity";
+import { Application } from "src/server/application/application.entity";
 
 export class CreateApplicationResponseDto {
   @ApiProperty()
