+++ packages/api/src/server/auth/auth.abstract.ts
