+++ packages/api/src/server/auth/service/auth.service.ts
@@ -1,7 +1,7 @@
 import { LoggerService } from "src/configuration/logger/logger.service";
 import { Injectable } from "@nestjs/common";
 import { JwtService } from "@nestjs/jwt";
-import { IAuthService, IAuthServicePayload } from "src/core/abstracts/auth.abstract";
+import { IAuthService, IAuthServicePayload } from "src/server/auth/auth.abstract";
 
 @Injectable()
 export class AuthService implements IAuthService {
