+++ packages/api/src/server/deployment-conncetion/service/deployment-connection.factory.ts
@@ -1,6 +1,9 @@
 import { Injectable } from "@nestjs/common";
 import { DeploymentConnection } from "src/repository/mongo/model";
-import { DeploymentConnectionDTO, UpdateDeploymentConnectionDTO } from "src/server/deployment-conncetion/deployment-connection.dto";
+import {
+  DeploymentConnectionDTO,
+  UpdateDeploymentConnectionDTO,
+} from "src/server/deployment-conncetion/deployment-connection.dto";
 
 @Injectable()
 export class DeploymentConnectionFactoryService {
