+++ packages/api/src/server/property/service/property.module.ts
@@ -1,5 +1,5 @@
 import { Module } from "@nestjs/common";
-import { DataServicesModule } from "../../../services/data-services/data-services.module";
+import { DataServicesModule } from "../../../repository/data-services.module";
 import { PropertyFactoryService } from "./property.factory";
 import { PropertyUseCases } from "./property.service";
