+++ packages/api/src/server/sse-services/event.entity.ts
