+++ packages/api/src/server/sse-services/service/sse-consume.module.ts
