+++ packages/api/src/server/sse-services/service/sse-consume.service.ts
